// src/api/generated/models/salePaymentMethod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type SalePaymentMethod = typeof SalePaymentMethod[keyof typeof SalePaymentMethod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SalePaymentMethod = {
  cash: 'cash',
  financed: 'financed',
  direct_installments: 'direct_installments',
  exchange: 'exchange',
  other: 'other',
} as const;