// src/api/generated/models/saleStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

/**
 * @nullable
 */
export type SaleStatus = typeof SaleStatus[keyof typeof SaleStatus] | null;


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SaleStatus = {
  negotiating: 'negotiating',
  sold: 'sold',
  canceled: 'canceled',
} as const;